+++ Source/AirlinesReservationSystem/AirlinesReservationSystem/AirlinesReservationSystem.csproj
@@ -182,9 +182,11 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controllers\AirlinesController.cs" />
-    <Compile Include="IReponsitory\GenericReponsitory.cs" />
-    <Compile Include="IReponsitory\IGenericRepository.cs" />
+    <Compile Include="Areas\Admin\AdminAreaRegistration.cs" />
+    <Compile Include="Areas\Admin\Controllers\AirlinesController.cs" />
+    <Compile Include="Areas\Admin\Controllers\HomeController.cs" />
+    <Compile Include="Areas\Admin\IReponsitory\GenericReponsitory.cs" />
+    <Compile Include="Areas\Admin\IReponsitory\IGenericRepository.cs" />
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\Airline.cs" />
     <Compile Include="Models\Airport.cs" />
@@ -217,7 +219,7 @@
     <Compile Include="Models\SeatClass.cs" />
     <Compile Include="Models\SeatDetailByFlight.cs" />
     <Compile Include="Models\SeatNumber.cs" />
-    <Compile Include="ReponsitoryModel\ReponsitoryAirlines.cs" />
+    <Compile Include="Areas\Admin\ReponsitoryModel\ReponsitoryAirlines.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -231,6 +233,9 @@
     <Content Include="Content\Site.css" />
     <Content Include="Scripts\bootstrap.js" />
     <Content Include="Scripts\bootstrap.min.js" />
+    <Content Include="Areas\Admin\Views\web.config" />
+    <Content Include="Areas\Admin\Views\Shared\_Layout.cshtml" />
+    <Content Include="Areas\Admin\Views\Home\Index.cshtml" />
     <None Include="Scripts\jquery-3.3.1.intellisense.js" />
     <Content Include="Scripts\jquery-3.3.1.js" />
     <Content Include="Scripts\jquery-3.3.1.min.js" />
@@ -276,14 +281,15 @@
     <Content Include="Views\Manage\VerifyPhoneNumber.cshtml" />
     <Content Include="Views\Shared\Lockout.cshtml" />
     <Content Include="Views\Shared\_LoginPartial.cshtml" />
-    <Content Include="Views\Airlines\Create.cshtml" />
-    <Content Include="Views\Airlines\Delete.cshtml" />
-    <Content Include="Views\Airlines\Details.cshtml" />
-    <Content Include="Views\Airlines\Edit.cshtml" />
-    <Content Include="Views\Airlines\Index.cshtml" />
+    <Content Include="Areas\Admin\Views\Airlines\Create.cshtml" />
+    <Content Include="Areas\Admin\Views\Airlines\Delete.cshtml" />
+    <Content Include="Areas\Admin\Views\Airlines\Details.cshtml" />
+    <Content Include="Areas\Admin\Views\Airlines\Edit.cshtml" />
+    <Content Include="Areas\Admin\Views\Airlines\Index.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
+    <Folder Include="Areas\Admin\Models\" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="fonts\glyphicons-halflings-regular.woff2" />
