+++ Source/AirlinesReservationSystem/AirlinesReservationSystem/Areas/Admin/IReponsitory/IGenericRepository.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AirlinesReservationSystem.IReponsitory
+{
+   public interface IGenericRepository<T> where T : class
+    {
+        IEnumerable<T> SelectAll();
+        T SelectById(object id);
+        void Insert(T obj);
+        void Update(T obj);
+        void Delete(object id);
+        void Save();
+    }
+}
