+++ Source/AirlinesReservationSystem/AirlinesReservationSystem/Areas/Admin/Views/Airlines/Delete.cshtml
@@ -1,0 +1,49 @@
+﻿@model AirlinesReservationSystem.Airline
+
+@{
+    ViewBag.Title = "Delete";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<h2>Delete</h2>
+
+<h3>Are you sure you want to delete this?</h3>
+<div>
+    <h4>Airline</h4>
+    <hr />
+    <dl class="dl-horizontal">
+        <dt>
+            @Html.DisplayNameFor(model => model.AirlineName)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.AirlineName)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.LogoImage)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.LogoImage)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.Slogan)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Slogan)
+        </dd>
+
+    </dl>
+
+    @using (Html.BeginForm()) {
+        @Html.AntiForgeryToken()
+
+        <div class="form-actions no-color">
+            <input type="submit" value="Delete" class="btn btn-default" /> |
+            @Html.ActionLink("Back to List", "Index")
+        </div>
+    }
+</div>
