+++ Source/AirlinesReservationSystem/AirlinesReservationSystem/Areas/Admin/Views/Home/Index.cshtml
@@ -1,0 +1,8 @@
+﻿
+@{
+    ViewBag.Title = "Index";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<h2>Index Admin</h2>
+
