--- Source/AirlinesReservationSystem/AirlinesReservationSystem/Views/Airlines/Details.cshtml
@@ -1,43 +1,0 @@
-﻿@model AirlinesReservationSystem.Airline
-
-@{
-    ViewBag.Title = "Details";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Details</h2>
-
-<div>
-    <h4>Airline</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.AirlineName)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.AirlineName)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.LogoImage)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.LogoImage)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.Slogan)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Slogan)
-        </dd>
-
-    </dl>
-</div>
-<p>
-    @Html.ActionLink("Edit", "Edit", new { id = Model.AirlineId }) |
-    @Html.ActionLink("Back to List", "Index")
-</p>
